+++ packages/jsActions/mobile-resources-native/src/camera/TakePictureAdvanced.ts
@@ -130,7 +130,7 @@ export async function TakePictureAdvanced(
 
     function takePicture(): Promise<ImagePickerV2Response | ImagePickerResponse | undefined> {
         return new Promise((resolve, reject) => {
-            const options = nativeVersionMajor === 2 ? getOptionsV2() : getOptionsV4();
+            const options = nativeVersionMajor === 2 ? getOptionsV2() : getOptions();
             getPictureMethod()
                 .then(method =>
                     method(options, (response: ImagePickerV2Response | ImagePickerResponse) => {
@@ -294,9 +294,10 @@ export async function TakePictureAdvanced(
         };
     }
 
-    function getOptionsV4(): CameraOptions | ImageLibraryOptions {
+    function getOptions(): CameraOptions | ImageLibraryOptions {
         const { maxWidth, maxHeight } = getPictureQuality();
         return {
+            presentationStyle: "fullScreen",
             mediaType: "photo" as const,
             maxWidth,
             maxHeight